+++ tinper-bee/bee-colorpicker/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="0.0.14"></a>
+## [0.0.14](https://github.com/tinper-bee/bee-colorpicker/compare/v0.0.13...v0.0.14) (2019-11-26)
+
+
+
 <a name="0.0.13"></a>
 ## [0.0.13](https://github.com/tinper-bee/bee-colorpicker/compare/v0.0.12...v0.0.13) (2019-08-30)