+++ tinper-bee/bee-input-number/CHANGELOG.md
@@ -1,5 +1,10 @@
-<a name="2.2.3"></a>
-## [2.2.3](https://github.com/tinper-bee/bee-input-number/compare/v2.2.2...v2.2.3) (2019-11-21)
+<a name="2.2.5"></a>
+## [2.2.5](https://github.com/tinper-bee/bee-input-number/compare/v2.2.4...v2.2.5) (2019-12-02)
+
+
+
+<a name="2.2.4"></a>
+## [2.2.4](https://github.com/tinper-bee/bee-input-number/compare/v2.2.2...v2.2.4) (2019-12-02)