+++ tinper-bee/bee-table/CHANGELOG.md
@@ -1,3 +1,28 @@
+<a name="2.2.21"></a>
+## [2.2.21](https://github.com/tinper-bee/bee-table/compare/v2.2.20...v2.2.21) (2019-12-05)
+
+
+
+<a name="2.2.20"></a>
+## [2.2.20](https://github.com/tinper-bee/bee-table/compare/v2.2.19...v2.2.20) (2019-12-04)
+
+
+
+<a name="2.2.19"></a>
+## [2.2.19](https://github.com/tinper-bee/bee-table/compare/v2.2.18...v2.2.19) (2019-11-30)
+
+
+
+<a name="2.2.18"></a>
+## [2.2.18](https://github.com/tinper-bee/bee-table/compare/v2.2.17...v2.2.18) (2019-11-25)
+
+
+
+<a name="2.2.17"></a>
+## [2.2.17](https://github.com/tinper-bee/bee-table/compare/v2.2.16...v2.2.17) (2019-11-14)
+
+
+
 <a name="2.2.16"></a>
 ## [2.2.16](https://github.com/tinper-bee/bee-table/compare/v2.2.15...v2.2.16) (2019-11-12)