+++ tinper-bee/bee-tooltip/changelog.md
@@ -1,3 +1,8 @@
+<a name="2.1.5"></a>
+## [2.1.5](https://github.com/tinper-bee/bee-tooltip/compare/v2.1.4...v2.1.5) (2019-12-06)
+
+
+
 <a name="2.1.4"></a>
 ## [2.1.4](https://github.com/tinper-bee/bee-tooltip/compare/v2.1.3...v2.1.4) (2019-09-23)